RoomToSpace: Fix an issue with some parameter names
For some reasons parameters with names like "ID_PRVKU" or "Obecný komentář" were not recognized when testing if parameter name was in parameters names set. Probably some character encoding issue but I was not able to figure out what was going on. Directly looking for parameter name in room/space parameters solved the issue.